fix(tui): name a paused child on the dock, and correct the bucketing rationale
Batched remediation for review round 2 (F1, F2, F3, F4, F5) and design round 2
(D4, D5). QA round 2 was PASS with nothing to fix.

D5 — a paused child and an ordinary cancel were pixel-identical rows with
OPPOSITE lifetimes. `pause` is mechanically a cancel, so both rendered `⊘
cancelled`; that told no lie while both expired, but the previous commit made
a pause persist and left the cancel expiring in five minutes, so the frame
started contradicting the intent the user had recorded. `status_glyph` gains a
`paused` branch — its own word, exactly as `interrupted` already has one and
for the identical stated reason. `⏸` is width-checked to one cell (the
neighbouring media-control marks are two-cell emoji, the trap GLYPH_INTERRUPTED
already documents). Threaded to the panel rows, the collapsed caption, the
eviction rank (a row that now outlives the window must not be evicted in
favour of a quiet success that is about to expire), the summary memo key (the
same trap `queued` documents: pausing changes the counts without changing
anything on the job row) and the full-page title. Reconciled with
`mobile/projection.py`, which has mapped this to `parked` all along — the
phone was showing a state the TUI did not.

F1/D4 — the `_bucketed_settle_stamp` docstring asserted a measurement the code
does not exhibit. Three streams failed to reproduce it; re-running it against
the real direction reproduces THEIRS, and I traced my figure to a probe that
advanced `settled_at` while holding `now` fixed — the inverse of the real
system, where flooring a frozen clock is a constant and the rejected variant
could not have coalesced. The code is unchanged and still correct; the claim
is replaced by the invariant `W mod b == 0`, which the 5% rule makes exactly
true (`W/b == 20`) on every shipped window. That is the honest argument: the
clock variant is accidentally correct here, while bucketing the stamp is
correct for any window.

NO POINT FIGURE for the coalescing is quoted anywhere now, deliberately —
mine, the reviewer's and QA's were all phase-dependent and all wrong as
general claims. Four agents, four numbers. A docstring should assert what
holds at every phase.

F3 — `test_a_batch_that_settled_together_leaves_together` passed with the
bucketing stubbed out, because it aged all six rows identically. Rewritten as a
300-phase sweep counting shrink events against an unbucketed control, and
mutation-tested: it now FAILS with the bucket neutralised. The `W mod b == 0`
invariant is pinned too, so the docstring's argument cannot silently rot.

F2 — the constant said "quantising DOWN", the function says "rounds UP"; the
code rounds up. F4 — one `float` window for both the predicate and the bucket
instead of `int` for one and `float` for the other. F5 — `_within_retention` is
`_within_roster_window`, matching the vocabulary the rule established.

Refs #524

+++ local_operator/harness/jobs.py
@@ -66,7 +66,7 @@ UNRESOLVED_TERMINAL_STATUSES = frozenset({"failed", "interrupted"})
 
 def retention_expired(
     job: Any,
-    retention_ms: int,
+    retention_ms: float,
     *,
     now: float | None = None,
     settled_at: float | None = None,
@@ -114,7 +114,7 @@ def retention_expired(
 
 def roster_expired(
     job: Any,
-    retention_ms: int,
+    retention_ms: float,
     *,
     paused: bool = False,
     now: float | None = None,
